+++ app/src/main/java/boulderstudios/com/boilerplate/data/errors/UserLoggedOutError.java
@@ -1,4 +1,4 @@
-package boulderstudios.com.boilerplate.errors;
+package boulderstudios.com.boilerplate.data.errors;
 
 /**
  * Created by yuval on 2/24/18.
