+++ app/src/main/java/boulderstudios/com/boilerplate/domain/interactor/IInteractor.java
@@ -1,10 +1,10 @@
-package boulderstudios.com.boilerplate.entities.interactor;
+package boulderstudios.com.boilerplate.domain.interactor;
 
 /**
  * Created by yuval on 2/13/18.
  */
 
-import boulderstudios.com.boilerplate.entities.model.Result;
+import boulderstudios.com.boilerplate.data.model.Result;
 
 /**
  * An interactor is equal to a "Use case". it's meant to encapsulate all the business logic of a
